Added schema concept for module.
