+++ source/abstracts/dynamics/module.hpp
@@ -14,54 +14,11 @@
 #define MODULE_ABSTRACT_HPP
 
 #include "common.hpp"
+#include "moduleschema.hpp"
 
 TEGRA_USING_NAMESPACE Tegra::Types;
 
 TEGRA_NAMESPACE_BEGIN(Tegra::Abstracts)
-
-/*!
- * \brief The ModuleType enum
- */
-__tegra_enum_class ModuleType : u8
-{
-    Index,      //!For global user service.
-    Admin,      //!For administrator service.
-    Root,       //!For root service.
-    System,     //!For system service.
-    Cron,       //!For cron job service.
-    Core,       //!For system core service.
-    Theme,      //!For system template.
-    Default,    //!For default global service.
-    Custom      //!For custom global service.
-};
-
-/*!
- * \brief The PermissionType enum
- */
-__tegra_enum_class PermissionType
-{
-    ReadOnly,
-    EditableOnly,
-    Restricted,
-    FullAccess,
-    ByService
-};
-
-/*!
- * \brief The ModuleInfo class
- */
-struct ModuleInfo __tegra_final
-{
-    Optional<u32>               codeName        {}; ///< A unique code for module.
-    OptionalString              name            {}; ///< A name for module.
-    OptionalString              description     {}; ///< A description for explanation module.
-    OptionalString              compiledDate    {}; ///< Compiled date for module.
-    Optional<SystemLicense>     license         {}; ///< License type for module.
-    ModuleType                  moduleType      {}; ///< The type of module.
-    Optional<SemanticVersion>   version         {}; ///< Version of module.
-    OptionalString              author          {}; ///< Author of module.
-    OptionalString              url             {}; ///< Url of module.
-};
 
 /*!
  * \brief The AbstractBaseClass class
