+++ source/abstracts/dynamics/moduleschema.cpp
@@ -1,0 +1,7 @@
+#include "moduleschema.hpp"
+
+TEGRA_USING_NAMESPACE Tegra;
+
+TEGRA_NAMESPACE_BEGIN(Tegra::Abstracts)
+
+TEGRA_NAMESPACE_END
